Move `getRecentlyUpdatedItemsAsync` to their own files
