+++ src/jsMain/kotlin/KtUniversalis.js.kt
@@ -46,34 +46,6 @@ internal actual val ktorClient = HttpClient(Js) {
 }
 
 /**
- * Returns the least-recently updated items on the specified world, along with the upload times for each item.
- * @param world The world to request data for.
- * @param entries The number of entries to return (default `50`, max `200`).
- * @throws UniversalisException The Universalis API returned an unexpected return code.
- */
-@JsExport @JsName("getLeastRecentlyUpdatedItemsByWorldAsync")
-fun getLeastRecentlyUpdatedItemsAsync(
-    world: World,
-    entries: Short? = null,
-) = GlobalScope.promise {
-    getLeastRecentlyUpdatedItems(world, entries)
-}
-
-/**
- * Returns the least-recently updated items on the specified world or data center, along with the upload times for each item.
- * @param dcName The data center to request data for.
- * @param entries The number of entries to return (default `50`, max `200`).
- * @throws UniversalisException The Universalis API returned an unexpected return code.
- */
-@JsExport @JsName("getLeastRecentlyUpdatedItemsByDataCenterAsync")
-fun getLeastRecentlyUpdatedItemsAsync(
-    dcName: DataCenter,
-    entries: Short? = null,
-) = GlobalScope.promise {
-    getLeastRecentlyUpdatedItems(dcName, entries)
-}
-
-/**
  * Returns the data currently shown on the market board for the requested item ID and world or data center.
  * @param worldDcRegion The world, data center, or region to retrieve data for. This may be an ID or a name. Regions should be specified as Japan, Europe, North-America, Oceania, China, or 中国.
  * @param itemId The item ID to retrieve data for.
@@ -198,34 +170,6 @@ fun getLeastRecentlyUpdatedItemsAsync(
  */
 @JsExport fun getMarketableItemsAsync() = GlobalScope.promise {
     getMarketableItems().toIntArray()
-}
-
-/**
- * Returns the most-recently updated items on the specified world, along with the upload times for each item.
- * @param world The world to request data for.
- * @param entries The number of entries to return (default `50`, max `200`).
- * @throws UniversalisException The Universalis API returned an unexpected return code.
- */
-@JsExport @JsName("getMostRecentlyUpdatedItemsByWorldAsync")
-fun getMostRecentlyUpdatedItemsAsync(
-    world: World,
-    entries: Short? = null,
-) = GlobalScope.promise {
-    getMostRecentlyUpdatedItems(world, entries)
-}
-
-/**
- * Returns the most-recently updated items on the specified data center, along with the upload times for each item.
- * @param dcName The data center to request data for.
- * @param entries The number of entries to return (default `50`, max `200`).
- * @throws UniversalisException The Universalis API returned an unexpected return code.
- */
-@JsExport @JsName("getMostRecentlyUpdatedItemsByDataCenterAsync")
-fun getMostRecentlyUpdatedItemsAsync(
-    dcName: DataCenter,
-    entries: Short? = null,
-) = GlobalScope.promise {
-    getMostRecentlyUpdatedItems(dcName, entries)
 }
 
 /**
